fix(sync): preserve liveness and protected Vitest ownership
